Rename command to `localllm`
The cli tool was originally called `llm` but this is already the name of
a pypi registered package/tool/project https://pypi.org/project/llm/.

This commit renames the tool from `llm` to `localllm`. It also supports
`locallm` because it is an easy typo to make (and me making it  - and
debugging the results - while making this change was actually the hardest
part of the change 🤦‍♀️🤦‍♀️🤦‍♀️), and in order to try to be
backwards compatible with anyone already using this, it also includes
aliasing `llm` to this tool as well - but only in the context of the
cloud workstations image.

For completeness we'd also register `localllm` at pypi but I'll hold off
until it is needed, since right now the focus of the repo is on using
the tool within cloud workstations.

Fixes #8