Rename `localllm` to `local-llm`
Hopefully the final rename! @mctavish suggested this name as a way to
avoid the typo issue (i.e. that localllm is really easy to typo as
locallm and so I was trying to support both to mitigate that) and after
having typed this incorrectly several times myself (and I found another
one while making this change XD) I am inclined to agree. It doesn't roll
off the fingers with the ease of `llm` but it's less error prone!